restructured layout to split posix and win32 implementations
